+++ Source/dvlnet/abstract_net.h
@@ -24,26 +24,17 @@ class abstract_net {
 public:
 	virtual int create(std::string addrstr, std::string passwd) = 0;
 	virtual int join(std::string addrstr, std::string passwd) = 0;
-	virtual bool SNetReceiveMessage(int *sender, void **data,
-	    uint32_t *size)
-	    = 0;
-	virtual bool SNetSendMessage(int dest, void *data,
-	    unsigned int size)
-	    = 0;
-	virtual bool SNetReceiveTurns(char **data, unsigned int *size,
-	    DWORD *status)
-	    = 0;
+	virtual bool SNetReceiveMessage(int *sender, void **data, uint32_t *size) = 0;
+	virtual bool SNetSendMessage(int dest, void *data, unsigned int size) = 0;
+	virtual bool SNetReceiveTurns(char **data, size_t *size, uint32_t *status) = 0;
 	virtual bool SNetSendTurn(char *data, unsigned int size) = 0;
 	virtual void SNetGetProviderCaps(struct _SNETCAPS *caps) = 0;
-	virtual bool SNetRegisterEventHandler(event_type evtype,
-	    SEVTHANDLER func)
-	    = 0;
-	virtual bool SNetUnregisterEventHandler(event_type evtype)
-	    = 0;
+	virtual bool SNetRegisterEventHandler(event_type evtype, SEVTHANDLER func) = 0;
+	virtual bool SNetUnregisterEventHandler(event_type evtype) = 0;
 	virtual bool SNetLeaveGame(int type) = 0;
-	virtual bool SNetDropPlayer(int playerid, DWORD flags) = 0;
-	virtual bool SNetGetOwnerTurnsWaiting(DWORD *turns) = 0;
-	virtual bool SNetGetTurnsInTransit(DWORD *turns) = 0;
+	virtual bool SNetDropPlayer(int playerid, uint32_t flags) = 0;
+	virtual bool SNetGetOwnerTurnsWaiting(uint32_t *turns) = 0;
+	virtual bool SNetGetTurnsInTransit(uint32_t *turns) = 0;
 	virtual void setup_gameinfo(buffer_t info) = 0;
 	virtual ~abstract_net() = default;
