+++ Source/dvlnet/base.h
@@ -19,17 +19,16 @@ public:
 
 	virtual bool SNetReceiveMessage(int *sender, void **data, uint32_t *size);
 	virtual bool SNetSendMessage(int playerId, void *data, unsigned int size);
-	virtual bool SNetReceiveTurns(char **data, unsigned int *size,
-	    DWORD *status);
+	virtual bool SNetReceiveTurns(char **data, size_t *size, uint32_t *status);
 	virtual bool SNetSendTurn(char *data, unsigned int size);
 	virtual void SNetGetProviderCaps(struct _SNETCAPS *caps);
 	virtual bool SNetRegisterEventHandler(event_type evtype,
 	    SEVTHANDLER func);
 	virtual bool SNetUnregisterEventHandler(event_type evtype);
 	virtual bool SNetLeaveGame(int type);
-	virtual bool SNetDropPlayer(int playerid, DWORD flags);
-	virtual bool SNetGetOwnerTurnsWaiting(DWORD *turns);
-	virtual bool SNetGetTurnsInTransit(DWORD *turns);
+	virtual bool SNetDropPlayer(int playerid, uint32_t flags);
+	virtual bool SNetGetOwnerTurnsWaiting(uint32_t *turns);
+	virtual bool SNetGetTurnsInTransit(uint32_t *turns);
 
 	virtual void poll() = 0;
 	virtual void send(packet &pkt) = 0;
